requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asyncpg==0.28.0
bcrypt==4.0.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.104.1
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
oauthlib==3.2.2
passlib==1.7.4
protobuf==4.25.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pydantic==2.4.2
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyparsing==3.1.1
python-dotenv==1.0.0
python-jose==3.3.0
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
uvicorn==0.23.2
watchfiles==0.21.0
websockets==12.0
